Move simulation test to separate package
